backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic-core==2.16.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
boto3==1.29.7
aioboto3==11.3.0
transformers==4.35.2
# PyTorch CPU-only for Windows Python 3.12 (install from PyTorch index)
# Install with: pip install torch==2.2.2+cpu torchvision==0.17.2+cpu torchaudio==2.2.2+cpu --index-url https://download.pytorch.org/whl/cpu
torch==2.2.2
spacy==3.7.2
# spaCy English model required: python -m spacy download en_core_web_sm
# Note: Model must be downloaded separately after installing spacy
sentence-transformers==2.2.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
regex==2023.10.3
dateparser==1.1.8  # Required for experience_parser.py date parsing
openai==1.3.0
faiss-cpu>=1.8.0  # Updated for Python 3.12 compatibility
scipy==1.11.4
nltk==3.8.1
textstat==0.7.3
faker==20.1.0
factory-boy==3.3.0
locust==2.17.0
moto==4.2.14
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

